refactor(ir): Split FlattenTile rewrite components
